service1/requirements.txt
flask==2.3.2
redis==5.0.1
openai-whisper==20230314
requests==2.31.0
gdown==4.7.1
